Further work on de-course-hardcoding refactoring.
This commit retcons some migrations; It may be necessary to run
$ python manage.py migrate Course 0008
$ python manage.py migrate
if you ran the versions of the new migrations (starting with 0009)
before this commit.

+++ Course/migrations/0009_course_tuwel_course_id.py
@@ -15,6 +15,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='course',
             name='tuwel_course_id',
-            field=models.PositiveSmallIntegerField(default=None, null=True),
+            field=models.PositiveIntegerField(default=None, null=True),
         ),
     ]
